refactor(bridge): split bot.go into 3 domain-specific files
bot.go was 1300+ lines mixing decision logic, agent crash alerts, and
jack notifications with the core event dispatch spine. Split into:
  - bot.go — struct, config, Run, event dispatch, slash commands, helpers
  - bot_decisions.go — NotifyDecision, UpdateDecision, DismissDecision,
    NotifyEscalation, modal handlers, block actions
  - bot_notifications.go — NotifyAgentCrash, NotifyJackOn/Off/Expired/Batch

No logic changes; pure code motion. Each file has the same package and
operates on *Bot methods.
